Continued removal of vue code from main compilation pathway
